- Refactor delete post in series

+++ src/Services/PostInSeries/PostInSeries.Api/Services/PostInSeriesService.cs
@@ -4,6 +4,7 @@ using PostInSeries.Api.Entities;
 using PostInSeries.Api.GrpcClients.Interfaces;
 using PostInSeries.Api.Repositories.Interfaces;
 using PostInSeries.Api.Services.Interfaces;
+using Shared.Constants;
 using Shared.Dtos.PostInSeries;
 using Shared.Dtos.Series;
 using Shared.Helpers;
@@ -56,7 +57,7 @@ public class PostInSeriesService(
         return result;
     }
 
-    public async Task<ApiResult<bool>> DeletePostToSeries(DeletePostInSeriesRequest request)
+    public async Task<ApiResult<bool>> DeletePostToSeries(Guid postId, Guid seriesId)
     {
         var result = new ApiResult<bool>();
         const string methodName = nameof(DeletePostToSeries);
@@ -64,19 +65,26 @@ public class PostInSeriesService(
         try
         {
             logger.Information("BEGIN {MethodName} - Deleting post with ID: {PostId} from series with ID: {SeriesId}",
-                methodName, request.PostId, request.SeriesId);
+                methodName, postId, seriesId);
 
-            var postInSeries = mapper.Map<PostInSeriesBase>(request);
-
+            var postInSeries = await postInSeriesRepository.GetPostInSeries(postId, seriesId);
+            if (postInSeries == null)
+            {
+                logger.Warning("{MethodName} - Post: {PostId} not found in series: {SeriesId}", methodName, postId, seriesId);
+                result.Messages.Add(ErrorMessagesConsts.PostInSeries.PostNotFoundInSeries);
+                result.Failure(StatusCodes.Status404NotFound, result.Messages);
+                return result;
+            }
+            
             await postInSeriesRepository.DeletePostToSeries(postInSeries);
             result.Success(true);
-            
+
             // Xoá cache
-            await cacheService.RemoveAsync(CacheKeyHelper.PostInSeries.GetAllPostInSeriesByIdKey(request.SeriesId));
+            await cacheService.RemoveAsync(CacheKeyHelper.PostInSeries.GetAllPostInSeriesByIdKey(seriesId));
 
             logger.Information(
                 "END {MethodName} - Successfully deleted post with ID: {PostId} from series with ID: {SeriesId}",
-                methodName, request.PostId, request.SeriesId);
+                methodName, postId, seriesId);
         }
         catch (Exception e)
         {
